extract IntUiBridgeCheckbox
